services/dt-pull-service/pyproject.toml
[tool.poetry]
name = "dt_pull_service"
version = "0.1.0"
description = "DT pull service for Catena-X"
authors = ["Dénes Surman (dddenes) <surmandenes@yahoo.com>"]

[tool.poetry.dependencies]
python-dotenv = "^1.0.1"
tractusx-sdk = "^0.0.4"
fastapi = "0.115.7"
httpx = "^0.28.1"
python = "3.12.3"
requests = "^2.32.3"
uvicorn = "0.30.3"

[tool.poetry.dev-dependencies]
pylint = "^3.3.6"
pytest = "^8.1.1"
pytest-cov = "^6.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint]
max-line-length = 120
max-args = 8
max-locals=20
disable = ["R0903", "R0917", "W0621", "W1203"]
